Remove unused method shouldCopyWhenCreatingSimilarTask()
Summary:
Nothing in the codebase calls or reads this function.

If it was ever used in some implementation of a custom field, then it was likely in context of the `Similar Task` functionality removed in 52f7446eeaa36ee1716fd1ed6e90b26b964e11ac.

Test Plan: Grep the codebase.

Reviewers: O1 Blessed Committers, avivey

Reviewed By: O1 Blessed Committers, avivey

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D25677