Phriction: fix issue with icons in table of contents on wide screens
Summary:
When using icons in front of a header title, a part of the icon was
not visible in the TOC on wide screens.

| Before    | After    |
|-----------|----------|
| {F2670860}|{F2670861}|

Look at the table of contents on the left side.

Closes T15920

Test Plan:
Steps to reproduce in a Phriction document:

```
== {icon users} How to Register-in ==

=== {icon heart spin} Credits ===
```

Verify if all icons are visible in TOC.

Reviewers: O1 Blessed Committers, aklapper, valerio.bozzolan

Reviewed By: O1 Blessed Committers, aklapper, valerio.bozzolan

Subscribers: aklapper, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15920

Differential Revision: https://we.phorge.it/D25854