Allow filtering tokens given by type of token
Summary:
Adds a new filter to the page /token/given/ to specify the type of token awarded.

Closes T15988

Test Plan:
Visited /token/ to see it works as before.

Specified a token type in the search menu, and checked that the results only include that token type.

Activate DarkConsole and inspect the Services tab to check the query with "burminate" has sense, and you see this:

    SELECT * FROM `token_given`   WHERE (tokenPHID IN ('PHID-TOKN-misc-4'))   ORDER BY `id` DESC LIMIT 101

Removed the token type in the search menu, the results still include everything, like before.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15988

Differential Revision: https://we.phorge.it/D25865