Update installation guide to point to GitHub mirrors
Summary: Cloning a repository from GitHub seems to be faster than on Phorge and also, the old Phabricator installation guide pointed to Phacility's GitHub mirrors. That's why this commit will update the installation guide to point to the GitHub mirrors of rP and rARC.

Test Plan: Run the git clone commands.

Reviewers: O1 Blessed Committers, avivey

Reviewed By: O1 Blessed Committers, avivey

Subscribers: avivey, aklapper, tobiaswiese, valerio.bozzolan, Matthew

Differential Revision: https://we.phorge.it/D25874