Expose time with File upload date
Summary:
Also show the time of a file uploaded, not only the date.
This is consistent for example with exposing the creation time of Maniphest tasks.

Closes T16030

Test Plan: Go to http://phorge.localhost/file/ while logged in and logged out, see that `Uploaded by testadmin on Wed, Apr 9` changed to `Uploaded by user on Wed, Apr 9, 10:48 AM`.

Reviewers: O1 Blessed Committers, mainframe98, valerio.bozzolan

Reviewed By: O1 Blessed Committers, mainframe98, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16030

Differential Revision: https://we.phorge.it/D25961