Fail more gracefully passing an array to AphrontRequest::getStr()
Summary:
It is possible to construct URIs with array parameters which end up in AphrontRequest::getStr() which previously just casted the passed value to a string via a `(string`) prefix.
This leads to an ugly `EXCEPTION: (RuntimeException) Array to string conversion`.

Instead, fail more gracefully with `(Exception) Value passed to "phutil_string_cast()" is an array; arrays can not be sensibly cast to strings.`

Closes T16102

Test Plan: Unclear.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16102

Differential Revision: https://we.phorge.it/D26077