When Points are disabled, replace "Point Limit" with "Count Limit" label
Summary:
When Maniphest points are disabled (the default) we have some labels talking about "Point Limit" but they are really working like a "Count Limit",
so we update the labels accordingly.

After this change, when the option maniphest.points is enabled:

- everything remains as-is

After this change, When that option is disabled (the default):

- labels about count limit, are "Count Limit"

We just avoid confusion with what really happens.

This task is NOT about changing the backend or changing behavior.

| Before     | After      |
|------------|------------|
| {F5082231} | {F5082234} |
| {F5083540} | {F5083543} |

Closes T16104

Test Plan:
Check the configuration page /config/edit/maniphest.points/ - have it empty (no points):

- visit a workboard column and click edit:  - you see "Count Limit" (no more Point Limit)
- visit the page /project/board/1/column/1/ - you see "Count Limit" (no more Point Limit)

To check for regressions, toggle the configuration /config/edit/maniphest.points/ to activate points:

- visit a workboard column and click edit:  - you see "Point Limit"
- visit the page /project/board/1/column/1/ - you see "Point Limit"

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, Matthew, Cigaryno

Maniphest Tasks: T16104

Differential Revision: https://we.phorge.it/D26078