Make Fulltext Search not crash when newResultObject() returns null
Summary:
Note that this is a forced crash. I was looking into Fulltext Search returning all and any objects also belonging to uninstalled applications, so I was experimenting with making `PhabricatorApplicationSearchEngine::newResultObject()` return `null` also when the corresponding application for the type of object is uninstalled (this method can already return `null`).
I believe it should not crash in this case.

Closes T16122

Test Plan:
1. Comment out code in `PhabricatorApplicationSearchEngine::newResultObject()` so this method will always return null (the current fallback)
2. Go to http://phorge.localhost/search/query/advanced/
3. Enter `bleh` in the `Query` field, hit the `Search` button
4. After applying the patch, get no crashes anymore (but incorrect search results not matching the entered search string but I consider that a separate problem, sorry)

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16122

Differential Revision: https://we.phorge.it/D26110