Diffusion Commit Search: Fix exception searching for ancestors without repository
Summary:
Do not crash with a vague `count(): Parameter must be an array or an object that implements Countable` exposed to the user when searching for ancestors but having no repository defined, but show the expected error message already defined in the code explaining to the user what to do.

Closes T16144

Test Plan: Go to http://phorge.localhost/diffusion/commit/query/advanced, set "Ancestors Of" to "1", do not set anything in "Repositories", search. Or see T16144 for details.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16144

Differential Revision: https://we.phorge.it/D26129