Remove transparency from workboard dropdown menus
Summary:
There is no obvious reason why on project workboards with a custom color, dropdown menu contents should be harder to read (due to transparency) than on workboards which do not have a custom color set.
As a side effect, this seems to solve another issue with dark mode in T15056.

Closes T16148

Test Plan: Go to a project workboard, "Change Background Color" in the "Manage" menu (cog icon) in the upper right corner, click on a dropdown button in a column header (pencil icon), see that the assignee icon of a task in a column is not visible behind the expanded dropdown menu.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16148

Differential Revision: https://we.phorge.it/D26135