Remove text-shadow from Aphront table headers
Summary:
In dark mode, the current `text-shadow` of table headers offers a nice fuzzy feeling and is a great alternative to the consumption of beverages.

We could either remove it completely or replace `white` with a new `{whitetext}` variable similar to the already existing `{blacktext}` variable. I prefer to keep it simple, so this patch removes the text-shadow.

Refs T15056

Test Plan:
* Go to http://phorge.localhost/settings/panel/display/ and set `Accessibility` to `Dark Mode (Experimental)`
* Go to http://phorge.localhost/settings/panel/editor/, scroll down to the section "External Editor URI Variables", look at the header text of the three columns.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26151