Hardcode text color of unexpected AphrontWebpageResponse output
Summary:
In dark mode, the current `text-shadow` of a potentially displayed unexpected response output offers a nice fuzzy feeling and is a great alternative to the consumption of beverages.

Variables like `{whitetext}` won't work here as it is a PHP file and not CSS. We could introduce a new CSS selector, or remove the text-shadow, or simply hardcode the currently implicit text color (taken from `body`) to black.

As there is also a custom background color gradient which does not change either between dark mode and default mode, let's just hardcode the text color.

Refs T15056

Test Plan:
* Go to http://phorge.localhost/settings/panel/display/ and set `Accessibility` to `Dark Mode (Experimental)`
* In `src/aphront/response/AphrontWebpageResponse::buildResponseString()`, override `$unexpected_output` with a custom string
* Go to http://phorge.localhost/ and look at the top of the page

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26152